Dispatch to allgather_packed_range from allgather for dynamic size types
This works currently by:
- Adding a `is_fixed_type` bool to StandardType classes
- doing SFINAE on Communicator::allgather based on `is_fixed_type`

This also adds default `typename Enable = void` template arguments to
`StandardType` and `Packing`. This is useful for enabling/disabling
specializations based on whether data members are fixed sizes or not.
I'm thinking of an example like `DualNumber<T,D>` where that class
may be fixed-size depending on whether D is fixed-size