Add Communicator::max(map)
Similar to broadcast(map) and sum(map), this is based on a shared
private Communicator::map_max() implementation so that it works for
both std::map and std::unordered_map.

There are two versions of Communicator::map_max(), one for fixed types
and one for non-fixed types. The SFINAE principal is used to control
which version gets called.