Create a Packing specialization for std::pair
This specialization is enabled when the std::pair is not of fixed
size, e.g. when either first or second is not of fixed size.
Size of first and second are computed using sizeof when
one is of fixed size, and using Packing::*size when not of fixed
size. I also introduce a couple of helper structs in order to
"automatically" determine the buffer_type for the Packing<pair>
specialization