Remove ifdef crud in parallel_sync
We'll rely on the lower-level communication routines to handle
differences between serial and parallel operation. This change required
adding a `NotADataType` overload for generic send of type `T` which is
only defined in serial. In parallel, it appears that we actually catch
all calls to `send` with more specific overloads such that we don't
need a definition of the generic `send`