Make DataType destructor virtual
As evidenced by the number of errors I get if I make this destructor
protected, we delete from `DataType` quite a bit. Consequently we have
to make sure that the `StandardType` destructor gets called because that
is generally what calls `free()`. This is ultimately what slew
idaholab/moose#19246, causing valgrind errors in
https://civet.inl.gov/job/880361/, and leading to reversion in
idaholab/moose#19331