Don't declare functions we can't define
We define these functions for particular specializations, but in each
case we have headers declaring those specializations and we don't need
the general declaration here.  Leaving the general declaration here just
makes it more confusing to debug when someone has the general header
included but not the particular specialized header(s) they need.