Instead, return an out_iter copy from unpack_range
This way we can still accept an lvalue temporary from code that wants to
pass us an lvalue temporary, but we can also hang on to a
post-incremented out_iter to correctly unpack multiple buffers to the
same output iterator.

This now requires changes in the code using unpack_range, but only if
the code was broken to begin with, so that's fine.