Use copy construction on out_iter, not assignment
We have users with non-assignable output iterator types; this fixes
backwards compatibility for them.  I don't like doing heap allocation
on a tiny little iterator here, but we only do it once for every time we
allocate up to 4MB of buffers on the heap so I'm sure it won't show up
in profiling.